feat: 도메인 클래스 구현
