+++ src/main/java/subway/StationRepository.java
@@ -2,5 +2,7 @@ package subway;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import subway.domain.Station;
+
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
