+++ src/main/java/subway/domain/SubwayLine.java
@@ -1,0 +1,34 @@
+package subway.domain;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class SubwayLine {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "SUBWAY_LINE_ID")
+	private Long id;
+
+	private String name;
+
+	private String color;
+
+	@OneToMany(mappedBy = "subwayLine")
+	private List<Station> stations = new ArrayList<>();
+
+}
