--- src/main/resources/application.properties
@@ -1,2 +1,0 @@
-spring.jpa.properties.hibernate.show_sql=true
-spring.jpa.properties.hibernate.format_sql=true
